Determine MAGICC version from MAGCFG_USER.CFG
MAGICC_VERSION env variable is no longer required
