+++ pymagicc/__init__.py
@@ -28,16 +28,7 @@ from ._version import get_versions
 __version__ = get_versions()["version"]
 del get_versions
 from .compat import get_param
-
-def _get_magicc_paths():
-    default_executable = os.path.join(
-        os.path.dirname(os.path.abspath(__file__)),
-        "MAGICC6/MAGICC6_4Download/magicc6.exe"
-    )
-
-    executable = os.environ.get('MAGICC_EXECUTABLE', default_executable)
-    return os.path.dirname(executable), os.path.basename(executable)
-
+from .paths import _get_magicc_paths
 
 _magiccpath, _magiccbinary = _get_magicc_paths()
